fix: 컬럼 변경사항 적용 - Influencer: nickname(unique constraint설정), profilePictureUrl(디폴트 길이 255 넘어가므로 길이 증가) - Media: graphMediaId(unique constraint설정), thumbnailUrl(디폴트 길이 255 넘어가므로 길이 증가) - Meta: Avg컬럼들 타입 Integer에서 Real(Float)으로 변경

+++ dp-server/src/main/java/com/cecd/dp/domain/meta/entity/Meta.java
@@ -21,10 +21,10 @@ public class Meta extends BaseEntity {
   private Long id;
 
   // 평균 좋아요 수
-  private Integer likeAvg;
+  private Float likeAvg;
 
   // 평균 댓글 수
-  private Integer commentsAvg;
+  private Float commentsAvg;
 
   // 팔로워 수
   private Integer followerCnt;
